requirements.txt
Django==6.0.2
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-filter==25.2
psycopg2-binary==2.9.11
redis>=5.0.0
